Fix tusd v2 hook compatibility, file size overflow, and add dedup
- Read hook type from body["Type"] and data from body["Event"] (tusd v2)
- Change match_files.size from Integer to BigInteger for large files
- Replace existing file when re-uploading same match_id + filename
- Add error handling to post-finish hook
- Document re-upload dedup behavior

+++ src/models.py
@@ -1,7 +1,7 @@
 import uuid
 from datetime import datetime
 
-from sqlalchemy import Boolean, DateTime, ForeignKey, Integer, JSON, String, Text, Uuid, func
+from sqlalchemy import BigInteger, Boolean, DateTime, ForeignKey, Integer, JSON, String, Text, Uuid, func
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from db import Base
@@ -137,7 +137,7 @@ class MatchFile(Base):
         Uuid, ForeignKey("matches.id", ondelete="CASCADE")
     )
     filename: Mapped[str] = mapped_column(String)
-    size: Mapped[int] = mapped_column(Integer)
+    size: Mapped[int] = mapped_column(BigInteger)
     tus_id: Mapped[str] = mapped_column(String, unique=True)
     uploaded_at: Mapped[datetime] = mapped_column(
         DateTime(timezone=True), server_default=func.now()
